test: add plan, subtask and task refs fixtures
