+++ miamoremustachio/ToDo/tests/fixtures/task_refs-data.js
@@ -1,0 +1,15 @@
+const PLAN_ID = '65c09742561b3a1051705b49';
+const USER_ID = '6561ff62413e98e914253b1f';
+const HEADERS = { Authorization: USER_ID };
+
+const VALID_TASK_REFS = [
+  '655449a15812e4abfc786f29',
+  '65544d9f5ae1fc8d350a6693',
+  '655451b35ae1fc8d350a6695',
+];
+
+module.exports = {
+  PLAN_ID,
+  HEADERS,
+  VALID_TASK_REFS,
+};
