+++ miamoremustachio/ToDo/tests/plans.test.js
@@ -4,21 +4,12 @@ const { app } = require('../app.js');
 const { database } = require('../modules/database/connection.js');
 const { getFinalResourceId } = require('./fixtures/test-utils.js');
 
-const USER_ID = '6561ff62413e98e914253b1f';
-const HEADERS = { Authorization: USER_ID };
-
-const VALID_PLAN = {
-  'title': 'escape from Strada',
-  'tasks': [
-    '655449a15812e4abfc786f29',
-    '65544d9f5ae1fc8d350a6693',
-    '655451b35ae1fc8d350a6695',
-  ],
-};
-
-const VALID_FIELDS = {
-  'title': 'become a true backender-shmackender',
-};
+const {
+  USER_ID,
+  HEADERS,
+  VALID_PLAN,
+  VALID_FIELDS,
+} = require('./fixtures/plan-data.js');
 
 let planId;
