+++ miamoremustachio/ToDo/tests/subtasks.test.js
@@ -4,22 +4,13 @@ const { app } = require('../app.js');
 const { database } = require('../modules/database/connection.js');
 const { getFinalResourceId } = require('./fixtures/test-utils.js');
 
-const TASK_ID = '65633d765d1d06dcad8ff7d0';
-const USER_ID = '6561ff62413e98e914253b1f';
-
-const VALID_SUBTASK = {
-  'title': 'drink korvalol',
-  'completed': false,
-};
-
-const VALID_FIELDS = {
-  'title': 'drink boyaryshnik',
-  'completed': true,
-};
-
-const INVALID_SUBTASK = {
-  'title': 'aa',
-};
+const {
+  TASK_ID,
+  HEADERS,
+  VALID_SUBTASK,
+  VALID_FIELDS,
+  INVALID_SUBTASK,
+} = require('./fixtures/subtask-data.js');
 
 let subtaskId;
 
@@ -31,7 +22,7 @@ describe('subtasks', () => {
   test('POST /tasks/:id/subtasks', async () => {
     const response = await request(app)
       .post(`/tasks/${TASK_ID}/subtasks`)
-      .set('Authorization', USER_ID)
+      .set(HEADERS)
       .send(VALID_SUBTASK);
     subtaskId = getFinalResourceId(response.text);
     expect(response.status).toBe(201);
@@ -40,7 +31,7 @@ describe('subtasks', () => {
   test('GET /tasks/:id/subtasks/:subtaskId', async () => {
     const response = await request(app)
       .get(`/tasks/${TASK_ID}/subtasks/${subtaskId}`)
-      .set('Authorization', USER_ID);
+      .set(HEADERS);
     expect(response.status).toBe(200);
     expect(response.body.title).toMatch(VALID_SUBTASK.title);
     expect(response.body.completed).toBeFalsy();
@@ -49,7 +40,7 @@ describe('subtasks', () => {
   test('PUT /tasks/:id/subtasks/:subtaskId', async () => {
     const response = await request(app)
       .put(`/tasks/${TASK_ID}/subtasks/${subtaskId}`)
-      .set('Authorization', USER_ID)
+      .set(HEADERS)
       .send(VALID_FIELDS);
     expect(response.status).toBe(200);
     expect(response.body.title).toMatch(VALID_FIELDS.title);
@@ -59,14 +50,14 @@ describe('subtasks', () => {
   test('DELETE /tasks/:id/subtasks/:subtaskId', async () => {
     const response = await request(app)
       .delete(`/tasks/${TASK_ID}/subtasks/${subtaskId}`)
-      .set('Authorization', USER_ID);
+      .set(HEADERS);
     expect(response.status).toBe(204);
   });
 
   test('GET /tasks/:id/subtasks', async () => {
     const response = await request(app)
       .get(`/tasks/${TASK_ID}/subtasks`)
-      .set('Authorization', USER_ID);
+      .set(HEADERS);
     expect(response.status).toBe(200);
     for (const subtask of response.body) {
       expect(subtask._id).not.toMatch(subtaskId);
@@ -78,7 +69,7 @@ describe('subtasks (invalid)', () => {
   test('POST /tasks/:id/subtasks', async () => {
     const response = await request(app)
       .post(`/tasks/${TASK_ID}/subtasks`)
-      .set('Authorization', USER_ID)
+      .set(HEADERS)
       .send(INVALID_SUBTASK);
     expect(response.status).toBe(400);
     console.log(`POST error message: \n${response.text}`);
@@ -89,7 +80,7 @@ describe('subtasks (not found)', () => {
   test('GET /tasks/:id/subtasks/:subtaskId', async () => {
     const response = await request(app)
       .get(`/tasks/${TASK_ID}/subtasks/${subtaskId}`)
-      .set('Authorization', USER_ID);
+      .set(HEADERS);
     expect(response.status).toBe(404);
   });
 });
